+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add CSV download functionality to the Sprunje tables UI.
 
 ## [6.0.0](https://github.com/userfrosting/theme-pink-cupcake/compare/6.0.0-rc.5...6.0.0) - 2026-06-12
 - No changes.